post,show,edit,destroy update

+++ app/Http/Controllers/TasksController.php
@@ -12,18 +12,34 @@ class TasksController extends Controller
     //getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        //メッセージ一覧を取得
-        $tasks = Task::all();
+        //認証済みの場合
+        $data = [];
+        if (\Auth::check()) {
+            //認証済みユーザを取得
+            $user = \Auth::user();
+            //ユーザのタスク一覧を取得
+            $tasks = $user->tasks();
+            //メッセージ一覧を取得
+            // $tasks = Task::all();
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+             return view('tasks.index', $data);
+        }
+        // メッセージ一覧ビューでそれを表示
+        return view('welcome', $data);
         
-        //メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+        //メッセージ一覧ビューでそれを表示(旧)
+        // // return view('tasks.index', [
+        //     'tasks' => $tasks,
+        //     ]);
     }
 
     //getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
     public function create()
     {
+        
         $task = new Task;
         
         //メッセージ作成ビューを表示
@@ -41,13 +57,18 @@ class TasksController extends Controller
             'content' => 'required|max:255', //20200806追加
             ]);
         //タスクを作成
-        $task = new Task;
-        $task->status = $request->status;    //20200807追加
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create ([
+            'status' => $request->status,    //20200824変更
+            'content' => $request->content,
+            ]);
+        //tasklist課題コード
+            // $task = new Task;
+            // $task->status = $request->status;    //20200807追加
+            // $task->content = $request->content;
+            // $task->save();
         
-        //トップページへリダイレクトさせる
-        return redirect('/');
+        //前のURLへリダイレクトさせる
+        return redirect('/tasks');
     }
 
     //getでtasks/idにアクセスされた場合の「取得表示処理」
@@ -57,9 +78,12 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
         
         //タスク詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
+         if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
             ]);
+        } 
+            return redirect('/tasks');
     }
 
     //getでtasks/id/editにアクセスされた場合の「更新画面処理表示」
@@ -69,9 +93,12 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
         
         //タスク編集ビューでそれを表示
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.edit', [
             'task' => $task,
             ]);
+        }
+        return redirect('/tasks');
     }
 
     //putでtasks/idにアクセスされた場合の「更新処理」
@@ -86,12 +113,13 @@ class TasksController extends Controller
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         //タスクを更新
+        if (\Auth::id() === $task->user_id) {
         $task->status = $request->status; //20200807追加
         $task->content = $request->content;
         $task->save();
-        
+        }
         //トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('/tasks');
     }
 
     //deleteでtasks/idにアクセスされた場合の「削除処理」
@@ -100,9 +128,11 @@ class TasksController extends Controller
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         //タスクを削除
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         
         //トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('/tasks');
     }
 }
